refactor(team): extract defaults + state normalization from broker.go
Phase 3 sixth slice. Pulls 12 default/normalization helpers into
broker_defaults.go:

  - defaultOfficeMembers / defaultOfficeMemberSlugs / defaultTeamChannels
    (load runtime manifest with company.DefaultManifest fallback)
  - repoRootForRuntimeDefaults (cwd or ".")
  - isDefaultChannelState / isDefaultOfficeMemberState
    (used by saveLocked's zero-state branch)
  - normalizeChannelSlug / normalizeActorSlug (slug canonicalization,
    preserving the __ DM separator)
  - ensureDefaultChannelsLocked / ensureDefaultOfficeMembersLocked
    (idempotent recovery; only seed when state is truly empty)
  - normalizeLoadedStateLocked (post-load fixup pass — dedup members,
    re-canonicalize slugs, reschedule request lifecycle)
  - reconcileOrphanedBlockedTasksLocked (one-shot migration for tasks
    blocked by since-terminated parents)

  - internal/team/broker.go: -318 LOC (2121 -> 1803)
  - internal/team/broker_defaults.go: +351 LOC (new)

Cumulative broker.go reduction across Phase 3:
  3404 -> 2859 (web proxy + sse)
       -> 2800 (auth + publish)
       -> 2700 (wiki lifecycle)
       -> 2376 (queries)
       -> 2121 (persistence)
       -> 1803 (defaults + normalization)
  Net: -1601 LOC; broker.go shrunk by 47%.

No behavior change. Tests pass.
